feat(realtime): Phase 4 - Realtime Layer with Socket.io and Redis pub/sub

+++ server/src/services/list.service.js
@@ -6,6 +6,8 @@ const Card = require('../models/Card');
 const { generateKeyBetween } = require('fractional-indexing');
 const logger = require('../utils/logger');
 
+const { emitToBoard } = require('../utils/emitter');
+
 /**
  * Creates a new list at the end of a board.
  * @param {string} boardId
@@ -21,6 +23,12 @@ async function createList(boardId, title) {
 
   const list = await List.create({ boardId, title, order });
   logger.info('List created', { listId: list._id, boardId });
+  
+  // NOTE: For newly created lists, initialize the cards array for the client
+  const listData = list.toObject();
+  listData.cards = [];
+  
+  emitToBoard(boardId, 'list_created', listData);
   return list;
 }
 
@@ -30,9 +38,17 @@ async function createList(boardId, title) {
  * @returns {Promise<void>}
  */
 async function deleteList(listId) {
+  const list = await List.findById(listId);
+  if (!list) return;
+
   await Card.deleteMany({ listId });
   await List.findByIdAndDelete(listId);
   logger.info('List deleted', { listId });
+  
+  emitToBoard(list.boardId.toString(), 'list_deleted', { 
+    listId, 
+    boardId: list.boardId 
+  });
 }
 
 module.exports = { createList, deleteList };
